refactor: move subcommands to pkg
