+++ pkg/cmd/completion/completion.go
