+++ pkg/cmd/kafka/create/create.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/spf13/cobra"
 
-	commonflags "github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhoas/cmdutil/flags"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhoas/kafka/flags"
+	commonflags "github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/cmdutil/flags"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/kafka/flags"
 	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/api/builders"
 	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/api/managedservices"
 )
