+++ pkg/cmd/kafka/credentials/credentials.go
