+++ pkg/cmd/kafka/status/status.go
