+++ pkg/cmd/kafka/topics/topics.go
@@ -3,10 +3,10 @@ package topics
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhoas/kafka/topics/create"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhoas/kafka/topics/delete"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhoas/kafka/topics/list"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhoas/kafka/topics/update"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/kafka/topics/create"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/kafka/topics/delete"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/kafka/topics/list"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/cmd/kafka/topics/update"
 )
 
 const (
