+++ pkg/cmd/logout/logout.go
